[Chore] SpotListFilterVM을 SpotListVM으로 통합 (#62)
